+++ src/test/parsec/flow-bundle-test.js
@@ -13,11 +13,11 @@ function testParser(parser, string) {
 }
 
 export default {
-    setUp: function(done) {
+    setUp: function (done) {
         done();
     },
-
-    'expect flatten result to be ok': function(test) {
+/*
+    'expect flatten result to be ok': function (test) {
         const string = 'foobar';
         // tests here
         const parser = C.char('f')
@@ -29,7 +29,7 @@ export default {
         test.done();
     },
 
-    'expect thenReturns to be ok when empty': function(test) {
+    'expect thenReturns to be ok when empty': function (test) {
         const string = 'some';
         // tests here
         const parser = F.any.rep().then(F.eos).thenReturns([]);
@@ -39,60 +39,196 @@ export default {
         test.done();
     },
 
-    'expect startsWith to start': function(test) {
+    'expect startsWith to start': function (test) {
         const string = ' world';
-        const object='hello';
+        const object = 'hello';
         // tests here
-        const parser = F.startsWith(object).then(C.string(' world')).then(F.eos.drop());
+        const parser = F.startsWith(object)
+            .then(C.string(' world'))
+            .then(F.eos.drop());
         testParser(parser, string);
         test.ok(accepted);
         test.equals(value.join(''), 'hello world');
         test.done();
     },
-    'test wordsUntil': function(test) {
+
+
+    'test moveUntilFast string': function (test) {
+        const line = Streams.ofString('soXYZso');
+
+
+        const combinator = F.moveUntil('XYZ');
+        const parser = combinator.parse(line);
+        const value = parser.value;
+        const offset = parser.offset;
+
+        test.equals(value, 'so');
+        test.equals(offset, 2);
+        test.done();
+    },
+    'test moveUntilFast string with continuation': function (test) {
+        const document = 'start-detect-XYZ-continues';
+        const line = Streams.ofString(document);
+
+        const start = C.string('start-');
+
+        const combinator = start.drop()
+            .then(F.moveUntil('XYZ'))
+            .then(C.string('XYZ-continues').drop())
+        const parser = combinator.parse(line);
+        const value = parser.value;
+        const offset = parser.offset;
+
+        test.equals(value, 'detect-');
+        test.equals(offset, document.length);
+        test.done();
+    },
+    'test moveUntilFast array of string with continuation': function (test) {
+        const document = 'start-detect-XYZ-continues';
+        const line = Streams.ofString(document);
+
+        const start = C.string('start-');
+
+        const combinator = start.drop()
+            .then(F.moveUntil(['ABC', 'ZE', 'XYZ']))
+            .then(C.string('XYZ-continues').drop());
+
+        const parser = combinator.parse(line);
+        const value = parser.value;
+        const offset = parser.offset;
+
+        test.equals(value, 'detect-');
+        test.equals(offset, document.length);
+        test.done();
+    },
+    'test moveUntilFast string fails': function (test) {
+        const document = 'start-detect-XYZ-continues';
+        const line = Streams.ofString(document);
+
+        const start = C.string('start-');
+
+        const combinator = start.drop()
+            .then(F.moveUntil('EEE'))
+            .then(C.string('XYZ-continues').drop());
+
+        const parsing = combinator.parse(line);
+
+        test.ok(!parsing.isAccepted());
+        test.done();
+    },
+    'test moveUntilFast array of string fails': function (test) {
+        const document = 'start-detect-XYZ-continues';
+        const line = Streams.ofString(document);
+
+        const start = C.string('start-');
+
+        const combinator = start.drop()
+            .then(F.moveUntil(['ABC', 'ZE', 'EEE']))
+            .then(C.string('XYZ-continues').drop());
+
+        const parsing = combinator.parse(line);
+
+        test.ok(!parsing.isAccepted());
+        test.done();
+    },
+    'test moveUntilFast fails if array stream': function (test) {
+        const document = ['More', 'XYZ'];
+        const line = Streams.ofArray(document);
+
+
+        const combinator = F.moveUntil(['ABC', 'ZE', 'XYZ']);
+        let found = false;
+        try {
+            combinator.parse(line);
+        } catch (e) {
+            if (e === 'Input source must be a String') {
+                found = true;
+            }
+        }
+
+        test.ok(found);
+        test.done();
+    },
+    'test moveUntilFastString fails if array stream': function (test) {
+        const document = ['More', 'XYZ'];
+        const line = Streams.ofArray(document);
+
+
+        const combinator = F.moveUntil('XYZ');
+        let found = false;
+        try {
+            combinator.parse(line);
+        } catch (e) {
+            if (e === 'Input source must be a String') {
+                found = true;
+            }
+        }
+
+        test.ok(found);
+        test.done();
+    },
+
+    'test moveUntil': function (test) {
         const line = Streams.ofString('I write until James appears');
 
-         
-        const combinator = F.moveUntil(C.string('James')).thenLeft(F.any);
+        const combinator = F.moveUntil(C.string('James')).then(F.any.drop());
         const value = combinator.parse(line).value;
 
         test.equals(value, 'I write until ');
         test.done();
     },
-    'test wordsUntil Not found': function(test) {
+    'test moveUntil Not found': function (test) {
         const line = Streams.ofString('I write until James appears');
-
 
         const combinator = F.moveUntil(C.string('Indiana'))
             .then(C.string('I'))
-            .thenLeft(F.any);
+            .then(F.any.drop());
         const accepted = combinator.parse(line).isAccepted();
 
         test.ok(!accepted);
         test.done();
     },
-    'test wordsUntil found with more parsers': function(test) {
+    'test moveUntil  found with failing parser': function (test) {
         const line = Streams.ofString('I write until James Bond appears');
 
+        const combinator = F.moveUntil(C.string('James'))
+            .then(F.dropTo(F.eos));
+        const accepted = combinator.parse(line).isAccepted();
+
+        test.ok(!accepted);
+        test.done();
+    },
+    'test dropTo with string': function (test) {
+        const line = Streams.ofString('I write until James Bond appears');
+
+        const combinator = F.dropTo('James')
+            .then(C.string(' Bond appears'))
+            .then(F.eos);
+        const accepted = combinator.parse(line).isAccepted();
+
+        test.ok(accepted);
+        test.done();
+    },
+    'test dropTo with parser': function (test) {
+        const line = Streams.ofString('I write until James Bond appears');
+
+        const combinator = F.dropTo(C.string('James'))
+            .then(C.string(' Bond appears'))
+            .then(F.eos);
+        const accepted = combinator.parse(line).isAccepted();
+
+        test.ok(accepted);
+        test.done();
+    },*/
+    'test moveUntil found with more parsers': function (test) {
+        const line = Streams.ofString('I write until James Bond appears');
 
         const combinator = F.moveUntil(C.string('James'))
-            .thenLeft(C.string('James Bond'))
-            .thenLeft(F.any);
+            .then(F.dropTo('appears'))
+            .then(F.eos.drop())
         const value = combinator.parse(line).value;
 
         test.equals(value, 'I write until ');
-        test.done();
-    },
-    'test wordsUntil  found with failing parser': function(test) {
-        const line = Streams.ofString('I write until James Bond appears');
-
-        
-        const combinator = F.moveUntil(C.string('James'))
-            .thenLeft(C.string('Indiana jones'))
-            .thenLeft(F.any);
-        const accepted = combinator.parse(line).isAccepted();
-
-        test.ok(!accepted);
         test.done();
     },
 };